refactor(#4712): pure-move relay recovery clusters
Split relay recovery planning, idle-tmux repair, action application, and tests into focused child modules so the production facade is no longer a grandfathered giant without changing behavior.
