+++ ARCHITECTURE.md
@@ -660,8 +660,12 @@ src/
 │   │   ├── relay_health/
 │   │   │   └── frontier.rs
 │   │   ├── relay_recovery/
-│   │   │   └── tests/
-│   │   │       └── circuit_breaker_apply.rs
+│   │   │   ├── tests/
+│   │   │   │   └── circuit_breaker_apply.rs
+│   │   │   ├── apply.rs
+│   │   │   ├── decision.rs
+│   │   │   ├── idle_tmux.rs
+│   │   │   └── tests.rs
 │   │   ├── router/
 │   │   │   ├── intake_dispatch/
 │   │   │   │   ├── attachment.rs
