+++ scripts/giant_file_registry.toml
@@ -734,20 +734,9 @@ deadline = "2026-08-31"
 decompose_issue = "#4712"
 
 # NOTE: src/services/discord/relay_recovery.rs previously had an [[entry]] here;
-# #3779 extracted auto-heal attempt accounting into a sibling module, dropping
-# the root below the prod-giant threshold. Its entry and ratchet baseline are
-# removed to avoid a ghost registration.
-
-[[entry]]
-# Re-crossed the 1000-prod-LoC threshold in #4030: watcher-cancel recovery now
-# routes the owning turn through the finalizer Cancel path, and the paired
-# FOREIGN-inflight stale cutoff is covered from the pending-start backstop. Keep
-# future relay recovery orchestration growth honest; split action apply/finalizer
-# glue into child modules once this P0 wedge recovery path stabilizes.
-file = "src/services/discord/relay_recovery.rs"
-owner = "discord-relay"
-deadline = "2026-08-31"
-decompose_issue = "#4712"
+# #4712 extracted decision, idle-tmux repair, and action apply clusters into
+# child modules, dropping the root below the prod-giant threshold. Its entry and
+# ratchet baseline are removed to avoid a ghost registration.
 
 [[entry]]
 # Crossed the 1000-prod-LoC threshold in #3739: worker-local LoopOwned Tokio
